Moved wrapper and dependency resolution at configuration time check to buildSrc
- Fixed bug in dependency resolution script displaying always the root project name
