+++ build.gradle
@@ -184,9 +184,9 @@ apply from: "gradle/dependencies.gradle"
 apply from: "gradle/dependenciesMetadataRules.gradle.kts"
 apply from: "gradle/minifiedDependencies.gradle"
 apply from: "gradle/testDependencies.gradle"
-apply from: "gradle/wrapper.gradle.kts"
+apply plugin: "configure-wrapper-tasks"
 apply plugin: "ide-configuration"
-apply from: "gradle/noDependencyResolutionDuringConfiguration.gradle.kts"
+apply plugin: "fail-dependency-resolution-during-configuration"
 
 // Calculate the set of released versions - do this at configuration time because we need this to create various tasks
 // At some point, make this a configuration rule
